pyproject.toml
[tool.poetry]
name = "east"
version = "0.1.0"
description = ""
authors = ["Shunta Komatsu <4620511@ed.tus.ac.jp>"]

[tool.poetry.dependencies]
python = "3.9.*"
fire = "^0.4.0"
lanms = "^1.0.2"
loguru = "^0.6.0"
numpy = "^1.22.2"
opencv-python = "^4.5.5"
pandas = "^1.4.1"
Pillow = "^9.0.1"
streamlit = "^1.5.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
gdown = "^4.3.1"
isort = "^5.10.1"
poethepoet = "^0.12.3"
pytest = "^7.0.1"

[tool.black]
line-length = 120

[tool.poe.tasks]
cuda113 = "pip install torch==1.10.2+cu113 torchvision==0.11.3+cu113 -f https://download.pytorch.org/whl/cu113/torch_stable.html"
demo = "streamlit run demo.py"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
